Move update_fits_wcs into ResampleStep.
